loc: move ilocation to namespace
