+++ BlasII.Randomizer.Map/Locations/ILocation.cs
